refactor: :recycle: Flowbite has been renamed to Pathogen
