+++ embedded_gems/pathogen/.erb_lint.yml
