+++ embedded_gems/pathogen/.erb_linters/linters.rb
