+++ embedded_gems/pathogen/.rubocop.yml
