--- test/components/previews/flowbite_button_preview/all_schemes.html.erb
@@ -1,13 +1,0 @@
-<div class="flex gap-1">
-  <%= render Flowbite::Button.new(scheme: :default) do %>
-    Button
-  <% end %>
-
-  <%= render Flowbite::Button.new(scheme: :primary) do %>
-    Button
-  <% end %>
-
-  <%= render Flowbite::Button.new(scheme: :danger) do %>
-    Button
-  <% end %>
-</div>
