--- test/components/previews/flowbite_button_preview/leading_visual.html.erb
@@ -1,4 +1,0 @@
-<%= render Flowbite::Button.new(scheme:, size:) do |button| %>
-  <% button.with_leading_visual_icon(icon: "magnifying-glass") %>
-  Search
-<% end %>
