--- test/components/previews/flowbite_button_preview/trailing_visual.html.erb
@@ -1,4 +1,0 @@
-<%= render Flowbite::Button.new(scheme:, size:) do |button| %>
-  <% button.with_trailing_visual_icon(icon: "arrow-right") %>
-  Proceed
-<% end %>
