+++ test/components/previews/pathogen_button_preview/all_schemes.html.erb
@@ -1,0 +1,13 @@
+<div class="flex gap-1">
+  <%= render Pathogen::Button.new(scheme: :default) do %>
+    Button
+  <% end %>
+
+  <%= render Pathogen::Button.new(scheme: :primary) do %>
+    Button
+  <% end %>
+
+  <%= render Pathogen::Button.new(scheme: :danger) do %>
+    Button
+  <% end %>
+</div>
