+++ test/components/previews/pathogen_button_preview/leading_visual.html.erb
@@ -1,0 +1,4 @@
+<%= render Pathogen::Button.new(scheme:, size:) do |button| %>
+  <% button.with_leading_visual_icon(icon: "magnifying-glass") %>
+  Search
+<% end %>
