+++ test/components/previews/pathogen_button_preview/trailing_visual.html.erb
@@ -1,0 +1,4 @@
+<%= render Pathogen::Button.new(scheme:, size:) do |button| %>
+  <% button.with_trailing_visual_icon(icon: "arrow-right") %>
+  Proceed
+<% end %>
