+++ test/components/previews/pathogen_button_preview/trailing_visual_svg.html.erb
@@ -1,0 +1,12 @@
+<%= render Pathogen::Button.new(scheme:, size:) do |button| %>
+  <% button.with_trailing_visual_svg do %>
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor">
+      <path
+        fill-rule="evenodd"
+        d="M3 10a.75.75 0 0 1 .75-.75h10.638L10.23 5.29a.75.75 0 1 1 1.04-1.08l5.5 5.25a.75.75 0 0 1 0 1.08l-5.5 5.25a.75.75 0 1 1-1.04-1.08l4.158-3.96H3.75A.75.75 0 0 1 3 10Z"
+        clip-rule="evenodd"
+      />
+    </svg>
+  <% end %>
+  Proceed
+<% end %>
